+++ lib/rails/auth/x509/filter/java.rb
@@ -1,23 +1,12 @@
-require "java"
-require "stringio"
-
 module Rails
   module Auth
     module X509
       module Filter
-        # Extract OpenSSL::X509::Certificates from Java's sun.security.x509.X509CertImpl
+        # Extract OpenSSL::X509::Certificates from java.security.cert.Certificate
         class Java
           def call(certs)
-            return unless certs
-            OpenSSL::X509::Certificate.new(extract_der(certs[0])).freeze
-          end
-
-          private
-
-          def extract_der(cert)
-            stringio = StringIO.new
-            cert.derEncode(stringio.to_outputstream)
-            stringio.string
+            return if certs.nil? || certs.empty?
+            OpenSSL::X509::Certificate.new(certs[0].get_encoded).freeze
           end
         end
       end